clip_text_to_image/requirements.txt
streamlit
transformers
torch
torchvision
Pillow
faiss-cpu
numpy